xe: jit: reorder: use copy plan for IR reorder
